requirements.txt
pywebview>=5.0
Pillow>=10.0
pyinstaller>=6.0
